pyproject.toml
[tool.poetry]
name = "thor-backend"
version = "0.1.0"
description = ""
authors = ["clogic <choiseungil29@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.1"
django-ninja = "^1.3.0"
black = "^24.8.0"
pyhumps = "^3.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
